pyproject.toml
[project]
name = "personal-team"
version = "0.1.0"
description = "Personal agentic team — AI agents on a Raspberry Pi"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "python-telegram-bot>=21.0",
    "apscheduler>=3.10.0",
    "python-dotenv>=1.0.0",
    "google-api-python-client>=2.100.0",
    "google-auth>=2.23.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.4.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
